+++ src/lib/import-as.js
@@ -45,5 +45,5 @@ export default importAs
  */
 /**
  * @suppress {nonStandardJsDocs}
- * @typedef {import('alamode').ÀLaModeReplacer} _alamode.ÀLaModeReplacer
+ * @typedef {import('alamode/types').ÀLaModeReplacer} _alamode.ÀLaModeReplacer
  */