+++ src/lib/rule.js
@@ -89,5 +89,5 @@ export { re }
  */
 /**
  * @suppress {nonStandardJsDocs}
- * @typedef {import('alamode').ÀLaModeReplacer} _alamode.ÀLaModeReplacer
+ * @typedef {import('alamode/types').ÀLaModeReplacer} _alamode.ÀLaModeReplacer
  */